requirements.txt
sentence-transformers
langchain
pypdf
pinecone
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain-groq
langchain_experimental
langchain_huggingface
fastapi
-e .